Refactor (#15)
The main changes are:

1. Removal of internal and external networks.
2. Abstraction over libp2p2 (cf. trait `Comm`) which allows using in-memory networks in tests (cf. `net::Star`).
3. `SailfishEvent` is split into `Message` and `Action`. Only `Message`s travel between nodes.
4. A `Coordinator` interfaces directly with the network and `Consensus`.
5. Renamed `Vote` to `Envelope` as we only treat timeout and no-vote messages as votes in `VoteAccumulator`s. `Envelope`s allow us to have signed commitments which we also use to exchange vertices between nodes.

The actual functionality is reduced compared to `main`, as the logic that already existed in `Consensus` (using vertex certificates and explicit votes) was removed. I hope to re-add the actual logic again next.

Tests currently make no progress as nothing kicks off the exchange of messages between nodes.