`is_valid` only requires the leader vertex ID.
Currently the node checking the validity of a vertex proposal requires
the leader vertex to be present in its local DAG in order to accept the
proposal with a strong edge to the leader vertex. This may cause issues
in case the node is falling behind by multiple rounds and upon resuming
regular traffic receives vertex proposals of rounds for which is has yet
to receive the leader vertex.

This changes checks that the proposed vertex has a strong path to the ID
of the leader vertex of the previous round without consulting its local
DAG, i.e. the proposal is accepted without the leader vertex itself
being known to the node.

NB that this will cause the proposed vertex to be buffered. Until the
node has received all vertices of the vertex edges, the proposal will
not be added to the local DAG.

This change is also meant to clarify the situation w.r.t. vertex IDs
which can be freely created by anyone.