First attempt to prune the DAG.
When we have delivered the vertices in a round `r` (in `order_vertices`)
we consider all vertices in round < `r` and remove those that are not
referenced in rounds >= `r`. We then also remove from the set of
delivered vertices those that are not found in the DAG anymore.

This approach fails if in future rounds a vertex arrives that references
other vertices in rounds < `r`. This could for example be a lagging
node. The paper assumes partial synchrony and depends on GST for
validity. We need to refine the conditions under which it is safe to do
the pruning and we also need to consider the set of buffered vertices.