Remove weak edges.
The variant of Sailfish we are implementing does not use weak edges.
This commit removes the `VertexId` and instead keeps only a set of
public keys as references. The round is always implicitely r - 1.

To aid with testing, the `Vertex` is made generic over the source, but
defaults to `PublicKey`. In the future this would allow us easier
generation of arbitrary DAGs for testing.