Always require evidence for rounds.
The following messages are received by a party:

1. Vertex proposals
2. Timeout messages
3. No-Vote messages
4. Timeout certificates

Currently only (4) is always proof of a quorum of parties. (1) is
only if a timeout or no-vote certificate is included. (2) and (3) never
contain any proof of a quorum.

In those cases where no quorum proof exists, parties can place the
message arbitrarily far into the future and the receiver has no good way
to decide if the message is malicious or a result of its own lagging in
the protocol.

To counteract this problem, this PR proposes to always require evidence
for the existence of a quorum. Every message is already indexed by a round
number and the evidence is required to apply to the predecessor of that
round (with the genesis round being the obvious exception).

Currently evidence is established as follows:

1. Timeout messages form a `Certificate<Timeout>`
2. No-Vote messages form a `Certificate<NoVote>`

This PR extends these two mechanisms by a third one:

3. Vertex proposals form a `Certificate<RoundNumber>`.

We now also require all messages to carry the evidence of the previous
round (except for the timeout certificate which is evidence of its own
round):

1. Vertex proposals have either a `Certificate<Timeout>` or a
   `Certificate<RoundNumber>`. In addition the existing logic for
   `Certificate<NoVote>` still applies.
2. Timeout messages likewise contain either a `Certificate<Timeout>` or
   a `Certificate<RoundNumber>` as evidence for the previous round.
3. No-Vote messages contain a `Certificate<Timeout>` as evidence for the
   previous round.

This enables the receiver to check that any message it receives is
either backed directly by a quorum or is no more than 1 round ahead of
such a backing.

To form a quorum from vertex proposals, the round number is signed by
the sender and added to a `VoteAggregator`, just like we currently do it
for timeouts and no-vote messages.