More careful timeout message sending.
Currently we send a timeout message on certain thresholds. The intention
was that we would do it only once and subsequent incoming votes would
change the number of votes. However this is not always the case, as we
also receive our own issued timeout messages which count as one vote.
This may result in a tight loop of sending and receiving our own timeout
messages.

To fix this we check that the message actually counted as a vote before
even comparing with the thresholds.