Add `Label` for logs.
Vertex edges use the public key of a node, as this uniquely identifies
the node itself. Our `NodeId`s on the other hand are separately assigned
numerical IDs. To help with associating nodes and DAG vertices this
commit adds a `Label` type which is the hash of the public key. The hash
implementation is not collision resistant, but for debug logs of small
sets this should not be a concern.

Based on the `Label` this commit also introduces `Dag::dbg_dag` to print
the DAG to stdout and `Vertex::dbg_edges` to print a vertex' round
number, source and parent vertices.