Move consensus logic into its own task.
A single consensus evaluation (e.g. to handle a message) takes a very
long time (> 1ms) and impedes the progress of the network which so far
runs in the same task.

This commit, spawns a separate task for consensus and uses
`tokio::task::block_in_pace` internally to run a consensus step.