Do not answer votes with more votes.
The problem is this: If nodes have delivered their message a lagging
node might continue to send votes because it has not received enough.
However if we just echo a vote with our own vote, this may result in
circular ping pong between nodes who have all delivered their message,
but think the other(s) have not. We need some ack to break this.