Fix some tests.
Some tests assume a certificate after the quorum number of vertex
proposals has been reached. However the key manager visits keys and
values in arbitrary order. The test logic assumes however that the
leader is part of the quorum subset, because otherwise consensus
must continue to run until the leader has been encountered resulting
in a vertex proposal for the next round with more that n - f edges.

While a test that checks these distinctions would be nice to have this
commit fixes the tests by using a `BTreeMap` which visits the keys in
sorted order and values in order of keys.