Always require evidence for rounds.
Sailfish parties exchange the following messages with each other:

1. Vertex proposals
2. Timeout messages
3. No-vote messages
4. Timeout certificates

Currently only (4) represents a proof of a quorum of parties. If (1)
includes a timeout or no-vote certificate for the preceding round,
then those are also proof of a quorum of parties. (2) and (3) on the
other hand never contain any such proof.

In those cases where no quorum proof exists, parties can place a
message arbitrarily far into the future and the receiver has no good
way to decide if the message is malicious or a result of its own
lagging.

We propose to require evidence for the existence of a quorum in every
message exchanged between parties. Every message is already indexed by
a round number and the evidence is required to apply to the same round
(in case of no-vote messages, and (trivially) for a timeout
certificate), or the immediate predecessor of that round (with the
genesis round being the obvious exception).

Currently evidence is established as follows:

- Timeout messages form a timeout certificate when at least n - f
  timeout messages have been received.
- No-Vote messages form a no-vote certificate when at least n - f
  timeout messages have been received.

We extend these two mechanisms with a third one:

- Vertex proposal messages form a round number certificate when at
  least n - f vertices have been received.

We require all messages to carry evidence as an extra field (except for
the timeout certificate which is evidence enough):

- Vertex proposals always include either a timeout or a round number
  certificate, and potentially a no-vote certificate if required by the
  existing protocol rules.
- Timeout messages likewise contain either a timeout or round number
  certificate as evidence for the preceding round.
- No-Vote messages contain a timeout certificate as evidence for the
  same round the no-vote applies to.

These changes enable a receiver to verify that any message it receives
is either backed directly by a quorum of parties or is no more than 1
round ahead of a quorum. It can thus conclude that a message that is
several rounds ahead is not malicious.

The round number certificate is derived from vertex proposal messages.
The vertex round number is now signed by the sender. This enables the
receiver to extract and aggregate these signed values until a
certificate can be formed (using the existing quorum threshold for
vertex proposals).