Allow DAG to retain committee size rounds.
Consider the following scenario: Node N receives its own proposals and
those of the other nodes, however its own outbound messages are
delayed. The rest of the committee moves on without N. Eventually, the
messages from N reach the others, but because N's proposals contain
edges to its own prior proposals the others can not add them unless
they have seen the previous ones from N. Since the committee might have
committed more than 2 rounds ahead of N, they may already have garbage
collected their DAG and therefore will not be able to resolve all of
N's proposals, causing N to be left behind as it continues to add its
own proposals which the others continue to reject.

This commit changes the GC to extend the DAG size to the full size of
the committee. This guarantees that each party is leader at least once
for the available DAG history and the others wait for the leader vertex
(or timeout) before committing.