RBC shutdown: Close channel immediately.
If RBC is shutdown and the caller does not consume messages anymore, the
worker may be blocked trying to deliver messages to the application and
does not detect the shutdown command.

This commit immediately closes the receiver, before sending the shutdown
command to the worker. The worker will either detect the shutdown upon
delivering a message or by receiving the command. In any case it
terminates operation.