Catchup when buffer is growing large.
Certain conditions might fill up the buffer of a node, e.g.

- It may start late and can not add vertices from other nodes because
  their edges can not be resolved in the DAG.
- Messages might not reach a node, likewise creating gaps which prevents
  moving vertices from buffer to DAG.

The general assumption behind the buffering is that eventually, all
vertices reach a node and they can be added to the DAG and for various
reasons this might not always be the case, causing the buffer to grow.

In here we replace our DAG with the first buffered round of vertices if
> 2f of vertices are available in that round, but only if the buffer has
already grown the size-of-committee rounds. If the first round has not
enough vertices, we drop it and check back upon the next incoming
vertex. This effectively limits the buffer size, but can cause a node to
skip over rounds if falling behind in order to catch up.