Rename `gc` to `cleanup` and add buffered vertices.
After we pruned the DAG and the buffer we add vertices from the lowest
buffer round to the DAG. This should help with cases where we are
unable to add vertices to the DAG because of varying delays before GST
where we end up in a situation that vertices arrive after we GC'd the
DAG and are unable to fully resolve vertex edges.

The change is assumed to be sage because we are at this point several
rounds ahead of the lowest buffer round and have at least a quorum of
vertices in our DAG in each of those rounds.