network: Fix max. message size handling.
In a weak moment with the help of autocompletion I selected the
`MAX_PAYLOAD_SIZE` constant as the limit for a message consisting of
potentially multiple frames. It makes no sense of course to limit
multiple frames to the payload size of a single noise payload.

This commit adds a constant `MAX_TOTAL_SIZE` (= 5MiB) to allow for
messages up to that value.