Check for quorum when advancing by NVC.
Since messages arrive out of order we may receive no-vote messages
before the vertex proposals arrive. When advancing based on a quorum of
no-vote messages we require that a quorum of vertices from the timeout
round have been received, otherwise we would create an invalid vertex
with not enough edges.