Update some comments.
Panic sections have been removed if the panic is not expected. In
accordance with the recommendation of `Option::expect`, expect messages
should state why a panic should not or can not happen. An informal proof
is ideal, but anything that convinces the reader that the `Option` is not
or should not be a `None` is fine too. If so, I do not think we should
include those in panic sections of doc comments. We also do not do this
for various other expressions that can fail, e.g. indexing with the
subscript operator.