Add `timeboost-types` and `timeboost-sequencer`.
Currently these crates live side by side with the existing timeboost
crates. `timeboost-types` contains a subset of `timeboost-core` and changes
the representation of transaction. It is meant to eventually replace
`timeboost-core`. It also does not make use of `SailfishBlock`, instead
it defines `CandidateList` and `InclusionList`.

This PR also removes the `Inbox` type from sailfish which was used like
a transaction or block queue. Queueing the data at this level makes it
hard to have accurate timestamps and to know when data is actually used
as payload in vertices. In its place sailfish consensus requires a
type implementing the new `DataSource` trait which can be asked for the
next data item. It allows the timeboost sequencer to construct a
`CandidateList` at the very moment sailfish wants to put it into a
newly created vertex. This change had an impact on tests, which so far
did not have to provide a data source.

The new `timeboost-sequencer` crate implements the inclusion phase based
on the payload data per round and creates scaffolding for the other
phases which need to be implemented. It is meant to be integrated with
the `timeboost` crate (eventually).