Detect multiple message proposals per round.
As pointed out in issue #276, a malicious party could attempt to send
multiple proposals per round and the current implementation creates a
tracker for each unknown proposal and votes on all of them.

This commit checks on each proposal that no two different digests for
the same message type and source exists in a round.