Require RBC votes to carry evidence.
Like with vertex proposals we want to prevent dishonest parties to send
us votes arbitrarily many rounds ahead, which we then need to store and
process. Therefore we require every vote to come with evidence taken
from the message this vote applies to, so we can be sure a quorum of
parties is behind the previous round. To avoid excessive signature checks
we only verify the evidence signatures for rounds where we do not already
have some data, as the evidence plays no role in RBC voting.