Refine type for RBC proposals to vertices.
Previously we would in various places accept `Message` as a type to
leave open the possibility that not only vertex proposals would be RBC
delivered, but also other kinds of messages. However with a more precise
type the code is simpler, so this commit allows only
`Envelope<Vertex<T>, _>` to be used for RBC proposals.

It also removes `instrument` tracing which often creates very long log
lines due to nesting.