Selectively re-send messages and votes.
Instead of re-broadcasting the message or our vote when revisiting the
status of messages, find out which voters are outstanding and send the
message or vote only to them. Spamming those who can only discard the
data serves no purpose.

Also exercise back-pressure on the network. If the capacity of the
channel to the application layer drops to 0 the worker waits. If the
application layer is only consuming from that channel after it has sent
this will result in a deadlock. Therefore we only accept inbound
messages from the network if the channel capacity is greater than 0.