Add explicit execution mode to sequencer.
Given the recently introduced `Action::Gc`, Sailfish can run ahead as
long as it does not garbage collect beyond the round last completed by
Timeboost. When the sequencer is passive, because the deduplication
cache is invalid, Sailfish should be allowed to proceed, because the
sequencer will not submit the data to subsequent phases. Only when
(re-)entering active mode does the GC cutoff round become relevant
again. In light of this, the execution logic is changed here to either
allow Sailfish to run unhindered, if in passive mode, or else to stop
execution when the GC round is greater than the last processed round.
The mode is determined by the outcome of the inclusion phase and is
derived from its cache inspection.