Use `CandidateListBytes` to estimate message size.
The network layer has a message size limit which requires the
sequencer to produce data items that do not cause the resulting vertex
message to exceed the allowed maximum. Since the message size is
dependent on the encoding format, this commit changes the payload data
type from `CandidateList` to `CandidateListBytes` (a newtype around
`Bytes`) and always serializes to bincode. When building the candidate
list it uses bincode's `SizeWriter` to calculate an estimate.