Serialize before handing data to the worker.
There are only two error cases for the worker to consider:

1. The serialized data is larger than what the network is willing to
   accept.
2. The overlay detects that the network is down.

To rule out (1) as much as possible, this commit moves the serialization
of outbound data out of the worker. This only leaves error case (2)
which is permament and so we can treat this error as fatal without
attempting to recover from it.