Add `RoundNumber` parameter to `Evidence::is_valid`.
To validate round number evidence it is always necessary to compare it
against the round number it is meant to support. To not forget this step
it is best to require it as an additional parameter. This also allows
for a better handling of genesis evidence which is valid for the genesis
round.