Use `Bucket`s in `Overlay`.
Due to message reordering the use of `SeqId` for garbage collection is
not correct. If for example a proposal in round `r + 1` arrived and we
send back a vote, yielding an `x : SeqId` and we later send out a
proposal for a round `r`, we get a `y : SeqId` with `y > x`. When
garbage collecting `r` we would remove all sequence numbers up to `y`
which includes `x` even though round `r + 1` is still active.

Here we change the approach and require senders to give us a bucket
number (usually corresponding directly to a round number) when sending a
message and also when garbage collecting.