Use a custom bounded channel for peers.
Since layers above `Network` have to cope with message loss they usually
perform retries (cf. `Overlay`), sending the same message until they
received acknowledgements from peers. If the peer is slow, this may
cause multiple identical copies of the message to appear in the queue
and subsequently sent over the network. This commit keeps only a single
message (as identified by its Id) in the channel. Either this message is
eventually sent out or the channel is dropped when it overflows or a
timeout occurs (just like currently).