Ensure certificates contain versioned data.
While `Envelope` and `Signature` are self-contained and can be verified
independently of a committee, a certificate that aggregates signatures
needs to ensure that they are all signed by keys of the same committee.
Consequently the commitment needs to include the version to prevent
adversaries from modifying the version which could influence certificate
verification, e.g. if the size of the committee changes between
versions.