Retry again in RBC worker.
In case a get request is either not answered or replied to with an
invalid response a different voter needs to be asked for the message.
Since the `Overlay` network does not know about this logic, the RBC
worker itself needs to periodically check messages with
`Status::Requested` and retry.