Tolerate `n - f` differences in tests.
With seed 15049687337786202439 the test `progress_after_random_prefix`
failed previously. Upon inspection the reason was, that node `A` was
waiting for vertex `(61, B)`, which the other nodes had received. `A`
was receiving other vertices, causing the quorum of committed rounds
and hence the GC cutoff point to move forward. By the time `A` received
`(61, B)` it had GC'ed `(58, A)` which the other parties included in
their outputs. This difference caused the test to fail.

This commit loosens the check to allow up to n - f differences in
outputs per round as per the safety properties we want to ensure. In the
future a more elaborate event analysis is possible since `A` only
differed in one round from the other parties.