Use `CommitteeVec` in `Producer`.
Due to different processing speeds, a node can receive messages from
another node that is already on a different committee, so the whole
committee selection by round number approach is flawed. In here we
always select a committee by ID and only use round numbers to determine
the time of the switch.