Fix race during committee change.
The round number `r` to switch committees is based on the
*committed* rounds of parties. Therefore it may happen
that a node has already broadcasted their proposal for
`r` as a member of the old committee before broadcasting
again for `r` as a member of the new committee. Naturally,
the digests differ, so a receiving node who hasn't switched
to the next committee yet, would reject the second proposal,
but as a special case, we chose to replace the old with the
new.

This means that theoretically between the time that the
next committee is known and the actual switch, a node may
vote multiple times for the same round, but never for the
same round and the same committee.