Introduce resharing bundle generation
Bundle generation happens outside of the Decrypter's Worker.
For resharing, this means that the Decrypter needs access to
the secret (obtained through the previous DKG/resharing).

To this end, the ThresholdEncKeyCell is moved, renamed and
now contains the full suite of threshold encryption key material.

Bundle generation logic has been modified to accommodate this
structural change in threshold encryption key state.