Make Worker aware of non-encrypted rounds
Committee switching lets the Sequencer schedule a future round for a switch.
In the decryption phase, a Worker only gets an inclusion list if it has
encrypted entries. Without it, the Worker may not realize it's off duty,
causing it to wait indefinitely. This change ensures all inclusion lists
are sent to Workers.