Fix arbitrary to 1.4.1 and add Cargo.lock
`arbitrary-v1.4.2` includes a module `details` behind the `derive`
feature flag. The various `alloy` crates depend directly on
`derive_arbitrary` and do not use the `derive` feature flag of
arbitrary. As a result, the `details` module in `arbitrary` will not be
visible, causing compilation errors.

To work around this, we add the `Cargo.lock` file with version 1.4.1 of
`arbitrary` and `derive_arbitary`.