Use externally provided block numbers.
The current way to have separate block number counters in nodes can not
provide consistent block numbers. If a counter misses just a single
increment, the node is permanently out of step with the others. We could
potentially do away with block numbers altogether and just deliver
certified blocks in the order they are given to us, however this
complicates evidence handling, catch-up and GC in certifiers.

The approach taken here is to require block numbers as input, i.e. as
part of the block protobuf file. The block provider needs to ensure that
block numbers are consistent, consecutive and strictly monotonically
increasing.