Ensure inclusion list round is monotonic.
After candidate lists of round `r` formed an inclusion list a delayed
candidate list for `r - 1` may be delivered by Sailfish. The debug
assertion --- while mostly true --- is not correct, because the delayed
candidate list might be the only delivery. This commit therefore removes
the assertion but also ensures that the round counter is monotonic.