Remove `run_dkg_handover_decryption_phase_e2e`.
This test is redundant since [PR #502][1] was merged and is often
failing on CI. The reason for its failure is a race condition: sometimes
not all nodes finish the resharing because of all the DKG requests they
send, less than t + 1 are answered because peers are themselves in an
incomplete DKG state or they are not in a running state. In any case they
are not replying. The test is then ultimately hanging and timing out
because it attempts to subsequently collect the inclusion list from all
decrypters and some can not provideit.

The test would need to be fixed by allowing enough time for the
decrypters of committee 1 to have their local DKG state ready before
the members of committee 2 send their DKG requests. For production it is
assumed that between scheduling the next committee and activating it,
sufficient time is given.

Rather than fixing the test this PR removes it because -- as mentioned
above -- it is redundant with another test.

[1]: https://github.com/EspressoSystems/timeboost/pull/502