Commit leader upon 2t+1 first messages of next round.
Rename the existing Sailfish coordinator event to `CoordinatorEvent` and
add an `Event` type to `sailfish_types` which encompasses `Message` and
a newly created `Info` type. The latter is used to notify the consensus
layer when >2t messages of round r+1 with an edge to the leader of r
have been received even without votes.